requirements.txt
# Core dependencies
anthropic==0.39.0
boto3==1.35.0
botocore==1.35.0

# HTTP and web scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
urllib3==2.2.3

# Data handling
python-dateutil==2.9.0
pytz==2024.1
pyyaml==6.0.2

# Optional dependencies
# RSS feed parsing
feedparser==6.0.11
